document and move route related funcs to router utils
