+++ packages/suite/src/components/suite/Link/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import NextLink, { LinkProps } from 'next/link';
 import { Link as TLink } from '@trezor/components';
-import { getPrefixedURL, isInternalRoute } from '@suite-utils/nextjs';
+import { getPrefixedURL, isInternalRoute } from '@suite-utils/router';
 
 interface Props extends LinkProps {
     className?: string;
