+++ packages/suite/src/components/wallet/TopNavigation/index.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import { colors, variables } from '@trezor/components';
 import { State } from '@suite-types/index';
 import { goto } from '@suite-actions/routerActions';
-import { getPrefixedURL } from '@suite-utils/nextjs';
+import { getPrefixedURL } from '@suite-utils/router';
 
 const { FONT_WEIGHT, FONT_SIZE, SCREEN_SIZE } = variables;
