+++ packages/suite/src/utils/suite/router.ts
@@ -1,22 +1,3 @@
-export const getApp = (url: string) => {
-    if (url === '/' || url.indexOf('/wallet') === 0) return 'wallet';
-    if (url.indexOf('/onboarding') === 0) return 'onboarding';
-
-    return 'unknown';
-};
-
-export const getParams = (url: string) => {
-    const split = url.split('#');
-    if (!split[1]) return {};
-    const parts = split[1].substr(1, split[1].length).split('/');
-
-    const params: { [key: string]: string } = {};
-    ['coin', 'accountId'].forEach((key, index) => {
-        params[key] = parts[index];
-    });
-    return params;
-};
-
 export interface Route {
     name: string;
     pattern: string;
@@ -66,6 +47,25 @@ export const routes: Route[] = [
     },
 ];
 
+export const getApp = (url: string) => {
+    if (url === '/' || url.indexOf('/wallet') === 0) return 'wallet';
+    if (url.indexOf('/onboarding') === 0) return 'onboarding';
+
+    return 'unknown';
+};
+
+export const getParams = (url: string) => {
+    const split = url.split('#');
+    if (!split[1]) return {};
+    const parts = split[1].substr(1, split[1].length).split('/');
+
+    const params: { [key: string]: string } = {};
+    ['coin', 'accountId'].forEach((key, index) => {
+        params[key] = parts[index];
+    });
+    return params;
+};
+
 export const getRoute = (name: string): string => {
     const entry = routes.find(r => r.name === name);
     if (!entry) {
@@ -74,4 +74,31 @@ export const getRoute = (name: string): string => {
         return '/';
     }
     return entry.pattern;
+};
+
+// Prefix a url with assetPrefix (eg. name of the branch in CI)
+// Useful with NextJS's Router.push() that accepts
+export const getPrefixedURL = (url: string) => {
+    const urlPrefix = process.env.assetPrefix || '';
+    return urlPrefix + url;
+};
+
+// Strips params delimited by a hashtag from the URL
+export const toInternalRoute = (url: string) => {
+    // chars before # belong to the first group, # and chars after to the optional second group
+    // eg. https://suite.corp.sldev.cz/wallet/account/#/eth/0 will be splitted to
+    // 'https://suite.corp.sldev.cz/wallet/account/' and '#/eth/0'
+    try {
+        const tokens = url.match(/([^#]*)(#.*)?/);
+        const group1 = tokens ? tokens[1] : url;
+        return group1;
+    } catch (error) {
+        console.error(error);
+        return url;
+    }
+};
+
+// Check if the URL/route points to an in-app page
+export const isInternalRoute = (route: string) => {
+    return routes.find(r => r.pattern === toInternalRoute(route));
 };
